No longer using time as argument to chaos
The time supplied by `simulate' is the length of one step, so it is
constant and it is useless as argument to `chaos'.

So I restored some function as they were before.

But using the coordinates of a cell as input to chaos is perfect.

If we do want to use the time elapsed, we'd need to use animate rather
than simulate.